[WORKSPACE, tree.BUILD] Use "tree" from head again, custom BUILD
In e1a2c900faa3cdb56cf7832d6a83a4b9bad6558c we pinned the dependency
on the "tree" repository to its version 0.1.6, which was the last
version to support Bazel. With this change, we port the last official
Bazel BUILD file to our own, private copy in `bazel/tree.BUILD`.

Our version of the BUILD file had to be adapted slightly, since the
standard http_archive repository rule we use to bring in the
dependency can only place a single BUILD file into the root of the
external repository.