[WORKSPACE, etc.] Make Python paths computed, and drop PY2 support.
Previously, the paths to the local Python installation were hardcoded
in bazel/python.BUILD, and had to be adjusted by the user to match
their local setup.

With this new system, the paths are computed by running the actual
Python interpreter and making it emit the relevant path.

We will only make this work for Python3, so Python2 support is removed
now.

As a consequence, the build inputs may change if the Python
interpreter changes; in that case users may need to run `bazel
clean`. Previously, changes in the path would have been changes of the
inputs and would automatically have caused rebuilding of the affected
targets.