[dmlab2d, python_system] Set wheel ABI, platform, interpreter tags.
This change augments the python_repo repository rule to collect and
expose the value of the system's Python interpreter's
"platform" (which contains OS+hardware platform), "interpreter" and
"wheel ABI tag" (which contains the CPython version). User code then
uses this data to set the relevant tags on the wheel-building target,
so that we can make different wheels for different platforms and
Python versions.

The "packaging" library is required locally to provide this
information.

For convenience, there is also a new wheel rule ":dmlab2d_wheel_local"
that produces a wheel file without ABI and platform constraints (i.e.
"none" and "any"). This wheel file will still work on the same
platform that was used to build it, and can be used as a fallback if
the above ABI and platform detection mechanism does not work.