[lib/env_lua_api, luajit.BUILD] Expose 'jit' support module
LuaJit includes a support module 'jit'. This change exposes that
module from the luajit sources, includes it in the dmlab2d runfiles
when luajit is used, and installs the relevant module searcher.

The latter is done unconditionally, as it does not affect anything
when the other Lua versions are used (the module directory simply
won't exist).