[game_scripts] Raise a helpful error if a color is missing from the palette.
The `if color then` check was passing for missing colors when the palate was specified as a read_settings table (the metatable for which specifies creating an empty table for any missing entries). This means that any palettes specified as settings would have vague "tensor length mismatch errors" if a color was missing from the settings.

Users have reported that missing colors have been unintentional and a source of errors.