[WORKSPACE, bazel/luajit.BUILD] Update LuaJit2 to v2.1 branch
We had previously pinned LuaJit2 to the v2.1.0-beta release, which is
the last names release and is from 2017. Now we follow the floating
"v2.1" upstream branch.

The new version contains some much-needed improvements for 64-bit
support, which are needed to make LuaJit2 compilable on MacOS (without
depending on deprecated workarounds).