[dmlab2d] On MacOS, remove unused symbols at link time.
This change adds the `-dead_strip` linker flag to the dmlab2d_pybind
module DSO build target, which causes LD64 to omit definitions of
unused symbols from the link. (This is the equivalent of -gc-sections
for the ELF linkers. For those linkers, Bazel's toolchain sets that
flag in opt mode by default, but for the Mach-O linkers neither the
built-in Bazel toolchain nor the bazelbuild/apple_support toolchain
sets -dead_strip, so we add it explicitly.)