[bazel/{png,luajit}.BUILD, bazelrc, precommit.yml] ARM64 support
Adds configuration cases to libpng and luajit to build for ARM64.

We also enable cross-compilation to allow building MacOS ARM64 binaries
from an x86_64 MacOS host by defining a new "target_arch" setting. This
setting is aliased by our bazelrc as "--macos_target_arch". The config
"macos_arm64" subsumes that target setting, so that config should work
both on native ARM64 hosts and for cross-compilation on x86_64 hosts.

Since the build results are not runnable on the GitHub x86_64 runners,
we only build, but do not run the unit tests for ARM64 in the precommit.