Build wheels and precommits with LuaJit and for MacOS/ARM64.
Affects [bazelrc, precommit, make_wheel, luajit.BUILD].

Building for ARM64 on the x86_64 GitHub action runners uses
cross-compilation. We add a new configurable setting to allow the
genrules that are executed on the host to be aware of the intended
target architecture, and we set this setting's value via a command
line flag subsumed by the "macos_arm64" config in our .bazelrc.

Since the build results are not runnable on the GitHub x86_64 runners,
we don't run the unit tests in those configurations.