[bazel/luajit.BUILD, bazelrc, precommit.yml] Allow internal unwinder.
We test this option by compiling with -fno-asynchronous-unwind-tables,
which removes the information necessary for the external unwinder to
work correctly. The external unwinder is more powerful, but not always
available (as this test case demonstrates), so it's useful to be able
to use the internal unwinder as an alternative.

The choice is gated by a configurable Bazel flag, which is aliased by
our .bazelrc as --luajit_external_unwinder. The default is "True";
setting it to "False" means we use the internal unwinder.

https://github.com/LuaJIT/LuaJIT/blob/224129a8e64bfa219d35cd03055bf03952f167f6/src/lj_err.c#L21-L85
explains the choice of unwinder in detail.

As a drive-by fix, we also remove the ineeffective linker flag `-u`,
which is not needed.